Add files via upload
This is the matrix implementation we talked about yesterday. I hope it is readable (I created the document in another program and then I converted it into excel format). You can see the progression from one matrix to the next. Have a look at the Table labelled 'N'. The 0th (uppermost) row and 0th (leftmost) column represent the colour of a vertex. In the beginning, all vertices have their own colour, e.g. vertex one has colour one, vertex 2 has colour 2 etc. The index of the row or column is equal to the index of the vertex. That is, vertex 3 is represented twice. Once in the third row and once in the the third column. Whenever two vertices are linked they will have a 0 in the cell that is in the row of the first vertex and the column of the second vertex. Whenever two vertices have the same colour they will get the same number in the 0th cell of their row and the 0th row of their column. 
The algorithm will progress through the entire matrix row by row, cell by cell. It will start at cell (1,2) that is the cell telling us that vertex one (row 1) and vertex 2 (column 2) have a connection. The connection is represented by the 0. The algorithm will go to the next cell (the one to the right) because it could not change the colour of the second vertex to colour 1. When the algorithm progressed to a cell where there is no value (represented by an underscore) then the algorithm will aim to change the colour of the vertex indicated by the column of the cell to the colour of the vertex indicated by the row of the cell. It can only do that when the vertex that it wants to recolour does NOT connect to another vertex that already has this colour. In order to check that the algorithm goes into the row of the target vertex and sees if the is any cell with a zero in it that has the same number (for the colour) in its column.
The frame around a cell in the matrix indicates that the vertex in the column of the highlighted cell will inherit the colour of the vertex in the row of the highlighted cell. The grey cells in the subsequent matrix indicate all the changes that have been made in response to that. Whenever a vertex shared its colour with another vertex it will get a -1 in the last cell of its row. This means that its colour can not be changed at a later time. The same goes for a vertex that has inherited its colour from another one. Lastly, When one row was scanned completely, the last entry in that row becomes a -1. This is because the vertex represented by the row has been processed in its entirety and does not have to be reconsidered at a later time. Note that once n-1 vertices have a '-1' the algorithm can stop. It does not mean that it reached the chromatic number but it means that it found the lowest upper bound that our algorithm is capable of. In the example shown in the excel sheet this number coincides with the chromatic number. 
You can see a draft of an algorithm in the textbook below. See if you can make sense of it or try to implement your own code. We are looking forward to seeing you again tomorrow. Get well soon!